chore: move pipeline component class to core pipeline module
