+++ wren-ai-service/src/core/pipeline.py
@@ -1,9 +1,13 @@
 import asyncio
 from abc import ABCMeta, abstractmethod
+from dataclasses import dataclass
 from typing import Any, Dict
 
 from hamilton.experimental.h_async import AsyncDriver
 from haystack import Pipeline
+
+from src.core.engine import Engine
+from src.core.provider import DocumentStoreProvider, EmbedderProvider, LLMProvider
 
 
 class BasicPipeline(metaclass=ABCMeta):
@@ -19,3 +23,11 @@ def async_validate(task: callable):
     result = asyncio.run(task())
     print(result)
     return result
+
+
+@dataclass
+class PipelineComponent:
+    llm_provider: LLMProvider = None
+    embedder_provider: EmbedderProvider = None
+    document_store_provider: DocumentStoreProvider = None
+    engine: Engine = None
